New Grip set correctly - Wrist set up
