+++ TeamCode/src/main/java/org/firstinspires/ftc/teamcode/UnitTesting/AutoTestRunDistance.java
@@ -1,4 +1,4 @@
-package org.firstinspires.ftc.teamcode.OpModes;
+package org.firstinspires.ftc.teamcode.UnitTesting;
 
 import com.qualcomm.robotcore.eventloop.opmode.Disabled;
 import com.qualcomm.robotcore.eventloop.opmode.LinearOpMode;
