+++ TeamCode/src/main/java/org/firstinspires/ftc/teamcode/UnitTesting/AutonomousTester.java
@@ -1,4 +1,4 @@
-package org.firstinspires.ftc.teamcode.autoRes.commands;
+package org.firstinspires.ftc.teamcode.UnitTesting;
 
 import com.qualcomm.robotcore.eventloop.opmode.Autonomous;
 import com.qualcomm.robotcore.eventloop.opmode.Disabled;
